% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi_sentinel.R
\name{ndvi_sentinel}
\alias{ndvi_sentinel}
\title{function to calculate NDVI from SAFE Sentinel folders}
\usage{
ndvi_sentinel(image_folder, clip_layer = NaN, export = FALSE)
}
\arguments{
\item{image_folder}{folder with the SAFE sentinel structure}

\item{clip_layer}{optional layer to clip the original image}

\item{export}{optional parameter to export the image in Tif format}
}
\value{
a stack object with the ndvi image
}
\description{
function to calculate NDVI from SAFE Sentinel folders
}
