% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel_low_tide_select.R
\name{sentinel_low_tide_select}
\alias{sentinel_low_tide_select}
\title{Select sentinel images that correspond to low tide periods and downloads it if required by the user}
\usage{
sentinel_low_tide_select(
  start_date,
  end_date,
  aoi,
  download = FALSE,
  download_path,
  login,
  pw,
  platform,
  product,
  sensor_id,
  time_window,
  country,
  site_id,
  cloud_percentage,
  tile_id = NULL
)
}
\arguments{
\item{start_date}{intial date in the format yyyy-mm-dd (e.g. 2018-01-01)}

\item{end_date}{final date in the format yyyy-mm-dd (e.g. 2018-12-01)}

\item{aoi}{shapefile with area of interest}

\item{download}{the user can decide to decide to download the images or not, default=FALSE}

\item{download_path}{path for the folder where images will be downloaded}

\item{login}{user name for the Copernicus SciHub account}

\item{pw}{password for the Copernicus SciHub account}

\item{platform}{Choice of Satellite (Sentinel-1, Sentinel-2, Sentinel-3)}

\item{product}{Choice of product type (L1C, L2Ap and L2A available)}

\item{sensor_id}{code to select both satellites (S2), only satellite A (S2A) and only satellite B (S2B)}

\item{time_window}{Time window (in hours) around low tide for selecting an image}

\item{country}{code to select tidal tables from a particular country, fr for France and all for all other countries}

\item{site_id}{numerical code to the tidal station, e.g. 114 for Le Croisic}

\item{cloud_percentage}{parameter to exclude images that have less than the percentage of clouds defined in this parameter}

\item{tile_id}{to select the tile code if needed}
}
\value{
a dataframe listing all the images that are available and/or downloaded
}
\description{
Requires a registred Copernicus SciHub account
}
