% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel_select.R
\name{sentinel_query}
\alias{sentinel_query}
\title{Function to query Sentinel Copernicus to check for available images}
\usage{
sentinel_query(start_date, end_date, aoi, download_path = "./", login,
  pw, platform, product, id, cloud_percentage)
}
\arguments{
\item{start_date}{intial date in the format yyyy-mm-dd (e.g. 2018-01-01)}

\item{end_date}{final date in the format yyyy-mm-dd (e.g. 2018-12-01)}

\item{aoi}{a shapefile with the area to be sampled}

\item{download_path}{path for the folder where the images should be downloaded}

\item{login}{user name for the Copernicus SciHub account}

\item{pw}{password for the Copernicus SciHub account}

\item{platform}{Choice of Satellite (Sentinel-1, Sentinel-2, Sentinel-3)}

\item{product}{Choice of product type (L1C, L2Ap and L2A available)}

\item{id}{code to select both satellites (S2), only satellite A (S2A) and only satellite B (S2B)}

\item{cloud_percentage}{parameter to exclude images that have less than the percentage of clouds defined in this parameter}
}
\value{
A dataframe with the list of files avilable for potential download
}
\description{
Requires a registred  Copernicus SciHub account (https://scihub.copernicus.eu/)
Uses the package getSpatialData to query and download images.
}
