% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_to_cds.R
\name{seurat_to_cds}
\alias{seurat_to_cds}
\title{seurat_to_cds}
\usage{
seurat_to_cds(
  seurat,
  assay = "SCT",
  preprocess = TRUE,
  dims = 50,
  norm = "none",
  import_umap = TRUE
)
}
\arguments{
\item{seurat}{The Seurat object to be converted.}

\item{assay}{Assay within the Seurat object to be imported.}

\item{preprocess}{Determines whether the new cds will be preprocessed. Default is TRUE.}

\item{dims}{Number of principal components returned after preprocessing.}

\item{norm}{Normalization method used in preprocess_cds. Default is "none".}

\item{import_umap}{Imports UMAP projection into the cell data set object. Default is TRUE.}
}
\value{
A cell data set (cds) object
}
\description{
Converts Seurat object to preprocessed cds.
}
\details{
Seurat to cell data set conversion function.
This function takes in a Seurat object and returns a cell data set object.
Automatically runs PCA on the new cell data set object.

I've only included this to stop warnings from document()
}
\examples{
new_cds <- seurat_to_cds(Seurat.object, assay = "SCT")
}
