% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal_height.R
\name{tidal_height}
\alias{tidal_height}
\title{internal funtion}
\usage{
tidal_height(H1, H2, T1, T2, t)
}
\arguments{
\item{H1}{height 1}

\item{H2}{height 2}

\item{T1}{time 1}

\item{T2}{time 2}

\item{t}{time point}
}
\value{
numeric value with tidal height
}
\description{
internal function
}
