% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal_tables.R
\name{tidal_tables}
\alias{tidal_tables}
\title{Function to recover tidal information between two dates}
\usage{
tidal_tables(start_date, end_date, site_id)
}
\arguments{
\item{start_date}{date in the format "yyyy-mm-dd" (e.g. "2018-12-01")}

\item{end_date}{date in the format "yyyy-mm-dd" (e.g. "2018-12-01")}

\item{site_id}{code for the tidal site (check the site url to find the code of the tidal station)}
}
\value{
a dataframe with the tide times corrected to UTC, the tidal height and
if they correspond to high tide (PM) or low tide (BM)
}
\description{
Function to recover tidal information between two dates. Read and parse tidal times from the site whttps://tides.mobilegeographics.com so an internet connection is needed
}
