% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tidal_heights.R
\name{tide_height}
\alias{tide_height}
\title{Function to predict the tidal height}
\usage{
tide_height(start_date, end_date, location = "all", site_id, user_time = NULL)
}
\arguments{
\item{start_date}{date in the format "yyyy-mm-dd" (e.g. "2018-12-01")}

\item{end_date}{date in the format "yyyy-mm-dd" (e.g. "2018-12-01")}

\item{location}{to chose the web site to get the tidal information, Fr,Ca or all}

\item{site_id}{code for the tidal site (check the site url to find the code of the tidal station)}

\item{user_time}{the specific time to determine the tidal height}
}
\value{
a dataframe with the tide times corrected to UTC, the tidal height and
if they correspond to high tide (PM) or low tide (BM)
}
\description{
Function to predict the tidal height in a specified site at a user specified time
}
