% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal_times_all.R
\name{tide_info_all}
\alias{tide_info_all}
\title{Function read and parse tidal times from the site ' https://tides.mobilegeographics.com}
\usage{
tide_info_all(tide_date, site_id)
}
\arguments{
\item{tide_date}{date in the format "yyyy-mm-dd" (e.g. "2018-12-01")}

\item{site_id}{code for the tidal site}
}
\value{
a dataframe with the tide times corrected to UTC, the tidal height and
if they correspond to high tide (PM) or low tide (BM)
}
\description{
read and parse tidal times from the site ' https://tides.mobilegeographics.com
}
