% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal_times_ca.R
\name{tide_info_ca}
\alias{tide_info_ca}
\title{Function read and parse tidal times from the site http://www.waterlevels.gc.ca}
\usage{
tide_info_ca(tide_date)
}
\arguments{
\item{tide_date}{date in the format yyyymmdd (e.g. 20181201)}
}
\value{
a dataframe with the tide times corrected to UTC, the tidal height and
if they correspond to high tide (PM) or low tide (BM)
}
\description{
read and parse tidal times from the site http://www.waterlevels.gc.ca
currently configured to read tides from Hopewell Cape (#170).
Uses the package rvest for the html parsing.
}
