% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal_times_fr.R
\name{tide_info_fr}
\alias{tide_info_fr}
\title{Function read and parse tidal times from the site http://maree.info/}
\usage{
tide_info_fr(tide_date, site_id = 114)
}
\arguments{
\item{tide_date}{date in the format yyyymmdd (e.g. 20181201)}

\item{site_id}{numerical code to the tidal station, e.g. 114 for Le Croisic}
}
\value{
a dataframe with the tide times corrected to UTC, the tidal height and
if they correspond to high tide (PM) or low tide (BM)
}
\description{
read and parse tidal times from the site http://maree.info/,
currently configured to read tides from Le croisic. Converts the time to UTC.
Uses the package rvest for the html parsing.
}
