\name{pkm}
\alias{pkm}
\alias{skm}
\title{Probabilistic K-Means}
\description{
  These functions were removed from \pkg{sentinel} and have been moved
  to the package \pkg{prokmeans} (currently on GitHub).
}
\usage{
pkm(...)
skm(...)
}
\arguments{
  \item{\dots}{(unused).}
}
\value{NULL}
\author{Emmanel Paradis}
\keyword{models}
